Network Gaming Bug Fixes - Actually Somewhat Playable Now
There were undiagnosed problems in GameUpdate and the related uses of it that were part of the problem in the use of the api.  Some of them were found out and solved, although a few big problems do remain.

+++ app/src/main/java/independent_study/fields/game/GameScreen.java
@@ -201,7 +201,8 @@ public class GameScreen extends Screen
         //playerSprite.destroy(); - Player Sprite Destruction Triggers This
         wallSpriteR.destroy();
         wallSpriteL.destroy();
-        obstacleSpriteManager.deleteAllObstacles();
+        if(obstacleSpriteManager != null)
+            obstacleSpriteManager.deleteAllObstacles();
         //fieldGame.clearGameScore();
     }
